requirements.txt
aiogram==2.25.1
selenium~=4.8.0
bs4~=0.0.1
beautifulsoup4~=4.11.1
python-dotenv~=0.21.1